requirements.txt
Django==4.0.1
django-cors-headers==3.10.1
django-daraja==1.2.0
djangorestframework==3.11.0
djongo==1.3.6
gunicorn==20.0.4
pymongo==3.12.1
python-decouple==3.6
python-dotenv==0.21.0
requests==2.28.1
sqlparse==0.2.4
urllib3==1.26.12
pika==1.3.1
